requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
pydantic==2.12.5
pydantic-settings==2.12.0
python-dotenv==1.2.1
python-multipart==0.0.22

anyio==4.12.1
starlette==0.50.0
typing_extensions==4.15.0

httpx==0.28.1
httpcore==1.0.9
h11==0.16.0
httptools==0.7.1
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0

timm==0.9.16
numpy==1.26.4
torch==2.1.2
torchvision==0.16.2
transformers==4.46.2
ultralytics==8.4.9

qdrant-client==1.16.2
grpcio==1.76.0
protobuf==6.33.5

pillow==12.1.0
opencv-python-headless==4.10.0.84

requests==2.32.5
PyYAML==6.0.3
orjson==3.11.7

huggingface-hub==0.36.0
tokenizers==0.20.3
safetensors==0.7.0
regex==2026.1.15
tqdm==4.67.2

aiofiles==24.1.0
filelock==3.20.3
fsspec==2026.1.0
packaging==26.0